# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "sneat",
        "description": "Description de votre API",
        "contact": {
            "name": "Nom du contact",
            "email": "contact@api.com"
        },
        "license": {
            "name": "Licence de l'API",
            "url": "URL de la licence"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/invitation/accept": {
            "post": {
                "tags": [
                    "Invitations"
                ],
                "summary": "accepte une demande d'amis qu on m'a envoyer",
                "description": "Accepte une demande d'amis",
                "operationId": "accepteEnAmis",
                "requestBody": {
                    "description": "ID de user a inviter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_invitation": {
                                        "type": "integer",
                                        "example": "id de invitation"
                                    },
                                    "friend_id": {
                                        "type": "integer",
                                        "example": "id de user qui m'a inviter"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "accepte user en amis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/invitation/invit-amis": {
            "post": {
                "tags": [
                    "Invitations"
                ],
                "summary": "Envoyer une demande d'amis",
                "description": "Envoyer une demande d'amis",
                "operationId": "inviteUserEnAmis",
                "requestBody": {
                    "description": "ID de user a inviter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_destinateur": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invitation user en amis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/invitation/joinGroup": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Rejoindre une groupe",
                "description": "Envoyer une invitation dans une groupe pour rejoindre ce groupe",
                "operationId": "joinGroup",
                "requestBody": {
                    "description": "ID de groupe",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_group": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rejoindre une groupe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/invitation/accepteJoinGroup": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Accepte un membre en groupe",
                "description": "Accepter une invitation de rejoindre dans un groupe , action fait par admin",
                "operationId": "accepteJoinGroup",
                "requestBody": {
                    "description": "ID de groupe",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_invitation": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "id_group": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Accepte un membre en groupe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/invitation/all-demande-groupe": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Tout les invitation en cours dans groupe",
                "description": "Invitation pas encore accepter pour rejoindre le groupe",
                "operationId": "allDemandeGroup",
                "requestBody": {
                    "description": "ID de groupe",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_group": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invitation en cours",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer votre_token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": [
                "read",
                "write"
            ]
        }
    ]
}